Move update.py to where it is used
